refactor: add changes to account for split 1 form and modals ticket ID 653
