+++ frontend/src/components/ui/button.tsx
@@ -19,6 +19,10 @@ const buttonVariants = cva(
         ghost:
           "hover:bg-accent hover:text-accent-foreground dark:hover:bg-accent/50",
         link: "text-primary underline-offset-4 hover:underline",
+        /** Primary brand action: light brand green that darkens on hover. */
+        brand: "bg-[#556830] text-white hover:bg-[#203622]",
+        /** Non-destructive warning (e.g. deactivate, archive). */
+        warning: "bg-orange-600 text-white hover:bg-orange-700",
       },
       size: {
         default: "h-9 px-4 py-2 has-[>svg]:px-3",
