+++ frontend/src/pages/admin/resident-profile/ResetPasswordDialog.tsx
@@ -3,15 +3,9 @@ import { Copy, Check } from 'lucide-react';
 import { toast } from 'sonner';
 import API from '@/api/api';
 import { User, ResetPasswordResponse, ServerResponseOne } from '@/types';
-import {
-    Dialog,
-    DialogContent,
-    DialogDescription,
-    DialogFooter,
-    DialogHeader,
-    DialogTitle
-} from '@/components/ui/dialog';
+import { DialogFooter } from '@/components/ui/dialog';
 import { Button } from '@/components/ui/button';
+import { FormModal } from '@/components/shared';
 
 type Phase = 'confirm' | 'result';
 
@@ -80,95 +74,91 @@ export function ResetPasswordDialog({
         }
     };
 
+    const isConfirmPhase = phase === 'confirm';
+
     return (
-        <Dialog open={open} onOpenChange={handleClose}>
-            <DialogContent>
-                {phase === 'confirm' ? (
-                    <>
-                        <DialogHeader>
-                            <DialogTitle>Reset Password</DialogTitle>
-                            <DialogDescription>
-                                Are you sure you want to reset{' '}
-                                {user.name_first} {user.name_last}'s password?
-                            </DialogDescription>
-                        </DialogHeader>
-                        <div className="py-4">
-                            <p className="text-sm text-gray-600">
-                                This will generate a temporary password for the
-                                resident. They will be required to change it on
-                                their next login.
-                            </p>
-                        </div>
-                        <DialogFooter>
-                            <Button
-                                variant="outline"
-                                onClick={() => handleClose(false)}
-                            >
-                                Cancel
-                            </Button>
-                            <Button
-                                onClick={() => void handleReset()}
-                                disabled={submitting}
-                                className="bg-[#556830] hover:bg-[#203622]"
-                            >
-                                Reset Password
-                            </Button>
-                        </DialogFooter>
-                    </>
-                ) : (
-                    <>
-                        <DialogHeader>
-                            <DialogTitle>Password Reset</DialogTitle>
-                            <DialogDescription>
-                                New temporary password for {user.name_first}{' '}
-                                {user.name_last}
-                            </DialogDescription>
-                        </DialogHeader>
-                        <div className="py-4">
-                            <div className="bg-gray-100 rounded-lg p-4 border border-gray-300">
-                                <div className="text-sm text-gray-600 mb-2">
-                                    Temporary Password
-                                </div>
-                                <div className="flex items-center gap-2">
-                                    <code className="flex-1 text-lg font-mono font-semibold text-[#203622]">
-                                        {tempPassword}
-                                    </code>
-                                    <Button
-                                        variant="outline"
-                                        size="sm"
-                                        onClick={() => void handleCopy()}
-                                    >
-                                        {copied ? (
-                                            <>
-                                                <Check className="size-4 mr-2" />
-                                                Copied!
-                                            </>
-                                        ) : (
-                                            <>
-                                                <Copy className="size-4 mr-2" />
-                                                Copy
-                                            </>
-                                        )}
-                                    </Button>
-                                </div>
+        <FormModal
+            open={open}
+            onOpenChange={handleClose}
+            title={isConfirmPhase ? 'Reset Password' : 'Password Reset'}
+            description={
+                isConfirmPhase
+                    ? `Are you sure you want to reset ${user.name_first} ${user.name_last}'s password?`
+                    : `New temporary password for ${user.name_first} ${user.name_last}`
+            }
+            titleClassName="text-foreground"
+        >
+            {isConfirmPhase ? (
+                <>
+                    <div className="py-4">
+                        <p className="text-sm text-gray-600">
+                            This will generate a temporary password for the
+                            resident. They will be required to change it on
+                            their next login.
+                        </p>
+                    </div>
+                    <DialogFooter>
+                        <Button
+                            variant="outline"
+                            onClick={() => handleClose(false)}
+                        >
+                            Cancel
+                        </Button>
+                        <Button
+                            onClick={() => void handleReset()}
+                            disabled={submitting}
+                            variant="brand"
+                        >
+                            Reset Password
+                        </Button>
+                    </DialogFooter>
+                </>
+            ) : (
+                <>
+                    <div className="py-4">
+                        <div className="bg-gray-100 rounded-lg p-4 border border-gray-300">
+                            <div className="text-sm text-gray-600 mb-2">
+                                Temporary Password
                             </div>
-                            <p className="text-sm text-gray-600 mt-4">
-                                Share this password securely with the resident.
-                                They will be prompted to change it on their next
-                                login.
-                            </p>
+                            <div className="flex items-center gap-2">
+                                <code className="flex-1 text-lg font-mono font-semibold text-[#203622]">
+                                    {tempPassword}
+                                </code>
+                                <Button
+                                    variant="outline"
+                                    size="sm"
+                                    onClick={() => void handleCopy()}
+                                >
+                                    {copied ? (
+                                        <>
+                                            <Check className="size-4 mr-2" />
+                                            Copied!
+                                        </>
+                                    ) : (
+                                        <>
+                                            <Copy className="size-4 mr-2" />
+                                            Copy
+                                        </>
+                                    )}
+                                </Button>
+                            </div>
                         </div>
-                        <DialogFooter>
-                            <Button
-                                onClick={() => handleClose(false)}
-                                className="bg-[#556830] hover:bg-[#203622]"
-                            >
-                                Done
-                            </Button>
-                        </DialogFooter>
-                    </>
-                )}
-            </DialogContent>
-        </Dialog>
+                        <p className="text-sm text-gray-600 mt-4">
+                            Share this password securely with the resident.
+                            They will be prompted to change it on their next
+                            login.
+                        </p>
+                    </div>
+                    <DialogFooter>
+                        <Button
+                            onClick={() => handleClose(false)}
+                            variant="brand"
+                        >
+                            Done
+                        </Button>
+                    </DialogFooter>
+                </>
+            )}
+        </FormModal>
     );
 }
